unified-ai-vendor/requirements.txt
flask
openai
google-generativeai
boto3
markdown
python-dotenv
